[REF] html_editor: add NoInlineRootPlugin
This commit adds the `NoInlineRootPlugin` plugin, extracting from the
selection plugin the logic that prevents inline content at the root
level into a dedicated plugin. This plugin ensures selections at the
editable root are corrected to prevent inline elements from being
inserted as direct children of the editable root.

task-4266746

X-original-commit: b429d37e20c712c5748a84c63c043e2b57fb92af
